+++ lizard_blockbox/management/commands/import_vertex_xls.py
@@ -35,12 +35,32 @@ class Command(BaseCommand):
 
     @transaction.commit_on_success
     def parse_sheet(self, sheet):
-        # First two rows are not vertexes.
-        row0 = sheet.row_values(1)[2:]
-        vertexes = []
-        for vertex in row0:
+        # The first row (0) of the sheet contains irrelevant comments;
+        # the first two columns are always location and reach. So
+        # we're interested in the headers in the rest of the 2nd row:
+        vertices = self.build_vertex_dict(sheet.row_values(1)[2:])
+
+        # Then, for every row after the first two, import it
+        for row_nr in xrange(2, sheet.nrows):
+            self.import_row(vertices, sheet.row_values(row_nr))
+
+    def build_vertex_dict(self, row_values):
+        vertices = []
+
+        for vertex in row_values:
             vertex = vertex.strip()
             header = ''
+            year = "2100"  # Let's use a default in case we don't find a year
+
+            # The first part of the vertex should be the year
+            if ':' in vertex:
+                parts = vertex.split(':')
+                first_part = parts[0].strip()
+                if first_part in models.VertexValue.YEARS:
+                    year = first_part
+                    vertex = ':'.join(parts[1:]).strip()
+
+            # Process the rest, which may contain a header
             if ':' in vertex:
                 # A vertex can contain multiple colons, only the
                 # first one is the header
@@ -50,21 +70,26 @@ class Command(BaseCommand):
                 vertex = ':'.join(text[1:]).strip()
             instance, _ = models.Vertex.objects.get_or_create(
                 header=header, name=vertex)
-            vertexes.append(instance)
 
-        vertexes = dict(enumerate(vertexes, 2))
+            instance.year = year  # Not saved on this model! But this
+                                  # is a convenient place to keep the
+                                  # variable around for below
 
-        for row_nr in xrange(2, sheet.nrows):
-            row = sheet.row_values(row_nr)
-            reach = models.Reach.objects.get(slug=row[1].strip())
-            riversegment, _ = models.RiverSegment.objects.get_or_create(
-                location=row[0], reach=reach)
+            vertices.append(instance)
 
-            for col_nr, vertex in vertexes.iteritems():
-                value = row[col_nr]
-                if not value:
-                    continue
-                models.VertexValue.objects.get_or_create(
-                    riversegment=riversegment,
-                    vertex=vertex,
-                    defaults={'value': value})
+        return dict(enumerate(vertices, 2))
+
+    def import_row(self, vertices, row):
+        reach = models.Reach.objects.get(slug=row[1].strip())
+        riversegment, _ = models.RiverSegment.objects.get_or_create(
+            location=row[0], reach=reach)
+
+        for col_nr, vertex in vertices.iteritems():
+            value = row[col_nr]
+            if not value:
+                return
+            models.VertexValue.objects.get_or_create(
+                riversegment=riversegment,
+                vertex=vertex,
+                year=vertex.year,  # Set the year we saved above
+                defaults={'value': value})