+++ CHANGES.rst
@@ -18,6 +18,10 @@ Changelog of lizard-blockbox
 
 - Removed unused ReferenceValue model.
 
+- Changed the import_vertex_xls script: headers may now start with the year (2050
+  or 2100) followed by a :, and this year value will be saved with the VertexValue.
+  If no year is present, assume 2100.
+
 
 0.6 (2013-03-28)
 ----------------